Conflicts:
	her/src/main/java/com/eeit95/her/controller/IdCatchServlet.java
	her/src/main/java/com/eeit95/her/controller/ShowPictureServlet.java
	her/src/main/java/com/eeit95/her/model/dao/font/FontDAOjdbc.java
	her/src/main/java/com/eeit95/her/model/font/AdvertisementBean.java
	her/src/main/java/com/eeit95/her/model/font/FontBean.java

+++ her/src/main/java/com/eeit95/her/model/card/CardTagInterface.java
@@ -6,10 +6,10 @@ import com.eeit95.her.model.tag.TagBean;
 
 public interface CardTagInterface {
 	public abstract List<CardTagBean> selectBycardId(String cardId);
-	public abstract List<CardTagBean> selectBytagId(short tagId);
+	public abstract List<CardTagBean> selectBytagId(int tagId);
 	public abstract List<CardTagBean> insert(List<CardTagBean> bean);
 	public abstract boolean deleteBycardId(String cardId);
-	public abstract boolean deleteBytagId(short tagId);
-	public abstract boolean delete(String cardId,short tagId);
+	public abstract boolean deleteBytagId(int tagId);
+	public abstract boolean delete(String cardId,int tagId);
 	public abstract List <CardTagBean> selectAll();
 }
